Added a sensor status monitoring skeleton, implemented elevation encoder
Sensor status routine is absorbed into sensor_monitor. This will require more implementation for other sensors but for now they will all be set to normal in the DB. I wrote an sql query to also grab the most recent sensor data to be used specifically for setting aspects of the project according to sensor statuses, specifically in diagnostics to set ERR when the elevation encoder has timed out. This feature will need to be tested on the telescope to ensure functionality, through some testing I was able to see that it will update the form if no check is performed and we just assume that the elevation sensor has timed out.