Fixed dat Sensor Monitoring Routine
OH YEAH IM SO HAPPY TO HAVE THIS DONE

But as far as stuff done, made the entire routine in sensor monitor inside of the rtcontroller not the control room controller. Removed some unused DB reads that were updated to pull from the RT data itself, this saves on time among other things. There will be another commit that removes unused code for the old weather monitoring routine. Additionally, I will further implement the sensor status conditions (this should be the last thing to finally put a stake through the heart of this problem).